+++ src/app/components/collection-text-types/metadata/metadata.component.ts
@@ -16,14 +16,14 @@ export class MetadataComponent implements OnInit {
   private collectionContentService = inject(CollectionContentService);
   private activeLocale = inject(LOCALE_ID);
 
-  readonly textItemID = input<string>('');
+  readonly publicationID = input.required<string>();
 
   loadingError$: Subject<boolean> = new Subject<boolean>();
   metadata$: Observable<any>;
 
   ngOnInit() {
     this.metadata$ = this.collectionContentService.getMetadata(
-      this.textItemID().split('_')[1] || '0', this.activeLocale
+      this.publicationID(), this.activeLocale
     ).pipe(
       catchError((error) => {
         console.error('Error loading metadata', error);