+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 - Migrate `@Output` custom events to the `output()` API.
 - Migrate `@Input` fields to the `input()` API.
 - Migrate decorator query fields to signal queries.
+- Refactor collection text id handling.
 - Deps: update `@angular/cli` to 20.2.0 and `@angular/core` to 20.2.1.
 - Deps: update `marked` to 16.2.0.
 - Deps (dev): update `@types/jasmine` to 5.1.9.